# ARROW-10402: [Rust] Refactor array equality