--- a/rust/arrow/src/array/array.rs
+++ b/rust/arrow/src/array/array.rs
@@ -50,7 +50,7 @@ const NANOSECONDS: i64 = 1_000_000_000;
 
 /// Trait for dealing with different types of array at runtime when the type of the
 /// array is not known in advance.
-pub trait Array: fmt::Debug + Send + Sync + ArrayEqual + JsonEqual {
+pub trait Array: fmt::Debug + Send + Sync + JsonEqual {
     /// Returns the array as [`Any`](std::any::Any) so that it can be
     /// downcasted to a specific implementation.
     ///
@@ -112,7 +112,7 @@ pub trait Array: fmt::Debug + Send + Sync + ArrayEqual + JsonEqual {
     /// // Make slice over the values [2, 3, 4]
     /// let array_slice = array.slice(1, 3);
     ///
-    /// assert!(array_slice.equals(&Int32Array::from(vec![2, 3, 4])));
+    /// assert_eq!(array_slice.as_ref(), &Int32Array::from(vec![2, 3, 4]));
     /// ```
     fn slice(&self, offset: usize, length: usize) -> ArrayRef {
         make_array(Arc::new(self.data_ref().as_ref().slice(offset, length)))
@@ -182,8 +182,7 @@ pub trait Array: fmt::Debug + Send + Sync + ArrayEqual + JsonEqual {
     /// assert_eq!(array.is_null(1), true);
     /// ```
     fn is_null(&self, index: usize) -> bool {
-        let data = self.data_ref();
-        data.is_null(data.offset() + index)
+        self.data().is_null(index)
     }
 
     /// Returns whether the element at `index` is not null.
@@ -200,8 +199,7 @@ pub trait Array: fmt::Debug + Send + Sync + ArrayEqual + JsonEqual {
     /// assert_eq!(array.is_valid(1), false);
     /// ```
     fn is_valid(&self, index: usize) -> bool {
-        let data = self.data_ref();
-        data.is_valid(data.offset() + index)
+        self.data().is_valid(index)
     }
 
     /// Returns the total number of null values in this array.
@@ -826,11 +824,6 @@ impl<T: ArrowPrimitiveType> From<ArrayDataRef> for PrimitiveArray<T> {
     }
 }
 
-/// Common operations for List types.
-pub trait ListArrayOps<OffsetSize: OffsetSizeTrait> {
-    fn value_offset_at(&self, i: usize) -> OffsetSize;
-}
-
 /// trait declaring an offset size, relevant for i32 vs i64 array types.
 pub trait OffsetSizeTrait: ArrowNativeType + Num + Ord {
     fn prefix() -> &'static str;
@@ -1003,14 +996,6 @@ impl<OffsetSize: OffsetSizeTrait> fmt::Debug for GenericListArray<OffsetSize> {
             fmt::Debug::fmt(&array.value(index), f)
         })?;
         write!(f, "]")
-    }
-}
-
-impl<OffsetSize: OffsetSizeTrait> ListArrayOps<OffsetSize>
-    for GenericListArray<OffsetSize>
-{
-    fn value_offset_at(&self, i: usize) -> OffsetSize {
-        self.value_offset_at(i)
     }
 }
 
@@ -1297,14 +1282,6 @@ impl<OffsetSize: BinaryOffsetSizeTrait> Array for GenericBinaryArray<OffsetSize>
     /// Returns the total number of bytes of memory occupied physically by this [$name].
     fn get_array_memory_size(&self) -> usize {
         self.data.get_array_memory_size() + mem::size_of_val(self)
-    }
-}
-
-impl<OffsetSize: BinaryOffsetSizeTrait> ListArrayOps<OffsetSize>
-    for GenericBinaryArray<OffsetSize>
-{
-    fn value_offset_at(&self, i: usize) -> OffsetSize {
-        self.value_offset_at(i)
     }
 }
 
@@ -1664,14 +1641,6 @@ impl<OffsetSize: StringOffsetSizeTrait> From<ArrayDataRef>
     }
 }
 
-impl<OffsetSize: StringOffsetSizeTrait> ListArrayOps<OffsetSize>
-    for GenericStringArray<OffsetSize>
-{
-    fn value_offset_at(&self, i: usize) -> OffsetSize {
-        self.value_offset_at(i)
-    }
-}
-
 /// An array where each element is a variable-sized sequence of bytes representing a string
 /// whose maximum length (in bytes) is represented by a i32.
 pub type StringArray = GenericStringArray<i32>;
@@ -1764,12 +1733,6 @@ impl FixedSizeBinaryArray {
     #[inline]
     fn value_offset_at(&self, i: usize) -> i32 {
         self.length * i as i32
-    }
-}
-
-impl ListArrayOps<i32> for FixedSizeBinaryArray {
-    fn value_offset_at(&self, i: usize) -> i32 {
-        self.value_offset_at(i)
     }
 }