… rust/arrow/src/array/equal/fixed_binary.rs
… rust/arrow/src/array/equal/fixed_list.rs
… rust/arrow/src/array/equal/list.rs
… rust/arrow/src/array/equal/mod.rs
… rust/arrow/src/array/equal/null.rs
… rust/arrow/src/array/equal/primitive.rs
… rust/arrow/src/array/equal/structure.rs
… rust/arrow/src/array/equal/utils.rs
… rust/arrow/src/array/equal/variable_size.rs
… rust/arrow/src/array/equal_json.rs
… rust/arrow/src/array/mod.rs
… rust/arrow/src/compute/kernels/concat.rs
… rust/arrow/src/compute/kernels/sort.rs
… rust/arrow/src/compute/kernels/take.rs